pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "uv-rental-skimmer"
version = "0.1.0"
description = "Fetches Utah Valley rental listings from Facebook Marketplace (via Bright Data), stores in SQLite, enriches with Claude, and publishes static HTML."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tomli>=2.0.0; python_version < \"3.11\"",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "anthropic>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "requests-mock>=1.11.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["uvrental*"]
